backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.12.1
redis==5.0.1
celery==5.3.4
langchain==0.0.350
langchain-groq==0.0.1
langchain-google-genai==0.0.8
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
firebase-admin==6.2.0
pdfplumber==0.10.0
python-docx==1.1.0
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0